Add effect timers and candy effect and improve brownie effect #12 #7
The effect of using candy has been added so it shrinks the player and resizes the player after the effect wears off. The double jump for the brownie effect has been fixed so the player can only jump twice at most when using it. Timers appear on screen when brownies and candies are used, a countdown from 10 happens and when it reaches 0 the effect stops.