Optimization push
Made a certain change which optimized the gameplay a lot more. To do with the particles. Particles if not being show will not render as much if they are far away. This ups the FPS and performance by a small amount but now we see the numbers on green to yellow rather than red. stat fps is on please turn off when pulled.